Extract PngEncoderLogic tests into its own test file.
